regular

+++ src/app/product-master/product-master.component.ts
@@ -13,6 +13,10 @@ import { Ng2CompleterModule } from 'ng2-completer';
 import { CompleterService, CompleterData } from 'ng2-completer';
 import { Crud } from '../crud';
 import { UnitService } from '../unit.service';
+import { randomInt } from 'crypto';
+import { HttpErrorService } from '../http-error.service';
+import { ErrorService } from '../error.service';
+import { MessageService } from '../message.service';
 
 @Component({
   selector: 'app-product-master',
@@ -28,57 +32,56 @@ export class ProductMasterComponent implements OnInit  {
    * field variables
    */
   id:any
-  primaryBarcode:string
-  itemCode:string
-  longDescription:string
-  shortDescription:string
-  ingredients:string
-  packSize:number
-  supplierName:string
-  departmentName:string
-  _className:string
-  subClassName:string
-  unitCostPrice:number
-  unitRetailPrice:number
-  profitMargin:number
-  standardUom:string
-  vat:number       
-  discount:number
-  quantity:number
-  maximumInventory:number
-  minimumInventory:number
-  defaultReOrderLevel:number
-  reOrderQuantity:number
+  primaryBarcode      : string
+  itemCode            : string
+  longDescription     : string
+  shortDescription    : string
+  ingredients         : string
+  packSize            : number
+  supplierName        : string
+  departmentName      : string
+  _className          : string
+  subClassName        : string
+  unitCostPrice       : number
+  unitRetailPrice     : number
+  profitMargin        : number
+  standardUom         : string
+  vat                 : number       
+  discount            : number
+  quantity            : number
+  maximumInventory    : number
+  minimumInventory    : number
+  defaultReOrderLevel : number
+  reOrderQuantity     : number
 
-  public items: string [] = []
-  public suppliers: string [] = []
-
-  public classNames: string [] = []
-  public departmentNames: string [] = []
+  public items           : string [] = []
+  public suppliers       : string [] = []
+  public classNames      : string [] = []
+  public departmentNames : string [] = []
 
   constructor(private httpClient: HttpClient ) {
-    this.id               ='';
-    this.primaryBarcode   ='';
-    this.itemCode         ='';
-    this.longDescription  ='';
-    this.shortDescription ='';
-    this.ingredients      ='';
-    this.packSize         =null;
-    this.supplierName     ='';
-    this.departmentName       ='';
-    this._className           ='';
-    this.subClassName         ='';
-    this.unitCostPrice    =null;
-    this.unitRetailPrice  =null;
-    this.profitMargin     =null;
-    this.standardUom      ='';
-    this.vat              =null;
-    this.discount         =null;
-    this.quantity         =null;
-    this.maximumInventory =null;
-    this.minimumInventory =null;
-    this.defaultReOrderLevel=null;
-    this.reOrderQuantity  =null;
+    this.id                  = '';
+    this.primaryBarcode      = '';
+    this.itemCode            = '';
+    this.longDescription     = '';
+    this.shortDescription    = '';
+    this.ingredients         = '';
+    this.packSize            = null;
+    this.supplierName        = '';
+    this.departmentName      = '';
+    this._className          = '';
+    this.subClassName        = '';
+    this.unitCostPrice       = null;
+    this.unitRetailPrice     = null;
+    this.profitMargin        = null;
+    this.standardUom         = '';
+    this.vat                 = null;
+    this.discount            = null;
+    this.quantity            = null;
+    this.maximumInventory    = null;
+    this.minimumInventory    = null;
+    this.defaultReOrderLevel = null;
+    this.reOrderQuantity     = null;
 
     
   }
@@ -106,9 +109,7 @@ export class ProductMasterComponent implements OnInit  {
         })
       }
     );
-
     this.departmentNames = await (new UnitService(this.httpClient)).getDepartmentNames()
-
    }
   
 
@@ -124,8 +125,8 @@ export class ProductMasterComponent implements OnInit  {
       shortDescription    : this.shortDescription,
       ingredients         : this.ingredients,
       packSize            : this.packSize,
-      supplier            :{supplierName:this.supplierName},
-      department          :{departmentName:this.departmentName},
+      supplier            : {supplierName : this.supplierName},
+      department          : {departmentName : this.departmentName},
       _class              : this._className,
       subClass            : this.subClassName,
       unitCostPrice       : this.unitCostPrice,
@@ -148,35 +149,34 @@ export class ProductMasterComponent implements OnInit  {
     /**
      * clear the fields
      */
-    this.id                  ='';
-    this.primaryBarcode      ='';
-    this.itemCode            ='';
-    this.longDescription     ='';
-    this.shortDescription    ='';
-    this.ingredients         ='';
-    this.packSize            =null;
-    this.supplierName            ='';
-    this.departmentName          ='';
-    this._className              ='';
-    this.subClassName            ='';
-    this.unitCostPrice       =null;
-    this.unitRetailPrice     =null;
-    this.profitMargin        =null;
-    this.standardUom         ='';
-    this.vat                 =null;
-    this.discount            =null;
-    this.quantity            =null;
-    this.maximumInventory    =null;
-    this.minimumInventory    =null;
-    this.defaultReOrderLevel =null;
-    this.reOrderQuantity     =null;
+    this.id                  = '';
+    this.primaryBarcode      = '';
+    this.itemCode            = '';
+    this.longDescription     = '';
+    this.shortDescription    = '';
+    this.ingredients         = '';
+    this.packSize            = null;
+    this.supplierName        = '';
+    this.departmentName      = '';
+    this._className          = '';
+    this.subClassName        = '';
+    this.unitCostPrice       = null;
+    this.unitRetailPrice     = null;
+    this.profitMargin        = null;
+    this.standardUom         = '';
+    this.vat                 = null;
+    this.discount            = null;
+    this.quantity            = null;
+    this.maximumInventory    = null;
+    this.minimumInventory    = null;
+    this.defaultReOrderLevel = null;
+    this.reOrderQuantity     = null;
   }
 
   showItem(item: object){
     
     /**
-     * render item information for display, these are displayed 
-     * automatically using two way data binding
+     * render item information for display, these are displayed automatically using two way data binding
      */
     this.id                   = item['id']
     this.primaryBarcode       = item['primaryBarcode']
@@ -196,17 +196,20 @@ export class ProductMasterComponent implements OnInit  {
     this.minimumInventory     = item['minimumInventory']
     this.defaultReOrderLevel  = item['defaultReOrderLevel']
     this.reOrderQuantity      = item['reOrderQuantity']
-
     this.supplierName         = item['supplier'].supplierName
     this.departmentName       = item['department'].departmentName
     this._className           = item['_class']
     this.subClassName         = item['subClass']
   }
 
+  generateItemCode(){
+    this.clear()
+    this.itemCode = "12345"
+  }
+
   validateData(){
     /**
-     * validates user inputs
-     * return false if validation fails
+     * validates user inputs return false if validation fails
      */
     var valid : boolean = true
     var message : string = ''
@@ -230,9 +233,29 @@ export class ProductMasterComponent implements OnInit  {
       valid = false
       message = message+'\nPack size is a required field!'
     }
+    if(isNaN(this.unitCostPrice) || this.unitCostPrice < 0 || this.unitCostPrice == null ){
+      valid = false
+      message = message+'\nInvalid cost price!'
+    }
+    if(isNaN(this.unitRetailPrice) || this.unitRetailPrice < 0 || this.unitRetailPrice == null ){
+      valid = false
+      message = message+'\nInvalid retail price!'
+    }
+    if(isNaN(this.profitMargin) || this.profitMargin < 0 || this.profitMargin == null ){
+      valid = false
+      message = message+'\nInvalid profit margin!'
+    }
+    if(isNaN(this.vat) || this.vat < 0 || this.vat == null ){
+      valid = false
+      message = message+'\nInvalid VAT!'
+    }
+    if(isNaN(this.discount) || this.discount < 0 || this.discount == null || this.discount >100){
+      valid = false
+      message = message+'\nInvalid discount!'
+    }
     
     if (message != ''){
-      alert('Error:'+message)
+      MessageService.showMessage('Error:'+message)
     }
     return valid
   }
@@ -244,18 +267,16 @@ export class ProductMasterComponent implements OnInit  {
      */
     if(this.validateData()==true){
       var item = this.getItemData()
-      //item['supplier_id'] = (new SupplierService(this.httpClient)).getSupplierId(null,this.supplier)
       if( this.id == '' || this.id == null || this.id == 0 ) {
         //save a new item
         this.httpClient.post(Data.baseUrl + "/items" , item)
         .subscribe(
           data=>{
-            window.alert('Item created successifully')
-            console.log(data)
+            MessageService.showMessage('Item created successifully')
             this.id=data['id']
           },
           error=>{
-            console.log(error)
+            ErrorService.showHttpError(error, 'Could not create item')
           }
         )
       } else {
@@ -263,12 +284,11 @@ export class ProductMasterComponent implements OnInit  {
         this.httpClient.put(Data.baseUrl + "/items/" + this.id , item)
         .subscribe(
           data=>{
-            window.alert('Item updated successifully')
-            console.log(data)
+            MessageService.showMessage('Item updated successifully')
             this.id=data['id']
           },
           error=>{
-            console.log(error)
+            ErrorService.showHttpError(error, 'Could not update item')
           }
         )
       }  
@@ -276,18 +296,21 @@ export class ProductMasterComponent implements OnInit  {
   }
   async searchItem() { 
     /**
-     * search item by id
-     * gets id from getItemId
+     * search item by id, gets id from getItemId
      */
+    if(this.primaryBarcode == '' && this.itemCode == '' && this.longDescription == ''){
+      MessageService.showMessage('Please enter a search key!')
+      this.clear()
+      return
+    }
     var itemId=''
     this.clear
     itemId = await (new ItemService(this.httpClient)).getItemId(this.primaryBarcode,this.itemCode,this.longDescription)
     if(itemId==''||itemId==null){
-      alert('No matching record')
+      //alert('Error: The requested record could not be found')
     }else{
-      var item
+      var item : any
       item =await (new ItemService(this.httpClient)).getItem(itemId)
-      console.log(item)
       this.showItem(item)
     }
   }
@@ -296,19 +319,25 @@ export class ProductMasterComponent implements OnInit  {
      * delete an item given its id
      */
     var id = this.id
-    this.httpClient.delete(Data.baseUrl+"/items/"+id)
-    .subscribe(
-      data=>{
-        console.log(data)
-        if(data==null){
-          this.clear()
-          alert('Item successifully deleted')
+    if(id == ''){
+      MessageService.showMessage('Please select item to delete!')
+      return
+    }
+    if(window.confirm('Delete the selected item?\nThe item will be removed and this action can not be undone.\nConfirm?')){
+      this.httpClient.delete(Data.baseUrl+"/items/"+id)
+      .subscribe(
+        data=>{
+          console.log(data)
+          if(data==null){
+            this.clear()
+            MessageService.showMessage('Item Successiful deleted')
+            alert()
+          }
+        },
+        error=>{
+          ErrorService.showHttpError(error, 'Could not delete item')
         }
-      },
-      error=>{
-        console.log(error)
-      }
-    )
+      )
+    }
   }
-
 }
